refactor: presentation 계층 관련 클래스 패키지 이동
