+++ src/main/java/roomescape/ui/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.ui.controller;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
@@ -17,8 +17,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.ReservationService;
 import roomescape.application.dto.response.ReservationResponse;
-import roomescape.auth.AuthenticationPrincipal;
-import roomescape.controller.dto.ReservationRequest;
+import roomescape.ui.controller.dto.ReservationRequest;
+import roomescape.ui.support.AuthenticationPrincipal;
 
 @RestController
 @RequestMapping("/reservations")
