+++ src/main/java/roomescape/ui/controller/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.ui.controller.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
