+++ src/main/java/roomescape/ui/controller/view/MemberViewController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.view;
+package roomescape.ui.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
